[v2.1][3/?] Differentiate compsci and probability options from their presets
Problem:
The 'compsci' and 'probability' options (true/false flags for only these specific tweaks) and the presets with the same name (including baseoptions, extraoptions, etc.) are currently clashing (latter overrides the former).

This is highly undesirable.

Resolution:
Rename single options to 'cs' and 'statistics', respectively, and keep 'compsci' and 'probability' naming for the presets.

Caveats:
This should not create breakage, but it does mean that this is a backwards-compatible functionality change, hence the new minor version 2.0 -> 2.1.