[v2.1][9/?] Rename 'hyperref' option to 'references'
Problem:
The 'hyperref' option does not only include the package 'hyperref' anymore ever since v2.

This is somewhat undesirable as it may cause confusion in the future.

Resolution:
Rename 'hyperref' option to 'references'.

Caveats:
This may create some breakage if any documents have used this option on its own, but that should be the minority (if there are even any). Due to this, it may not be a fully backwards-compatible functionality change, but in my opinion still minor enough to warrant only a minor version bump 2.0 -> 2.1.