[v2.2] cs -> computerscience, statistics -> probabilitytheory
Problem:
The 'cs' and 'statistics' options are currently named very badly: 'cs' is not in line with the fact that options should not be abbreviated and 'statistics' is plainly a misnomer

Especially the second case is not desirable.

Resolution:
Rename options to 'computerscience' and 'probabilitytheory', respectively, making sure to keep avoiding clashes with 'compsci' and 'probability' naming for the presets.

Caveats:
This may create some breakage if any documents have used these options on their own. Due to this, it may not be a fully backwards-compatible functionality change, but in my opinion still minor enough to warrant only a minor version bump 2.1 -> 2.2.